Move returnProblemsTable from common.tsx to clinicalInfoService.tsx. Create unit test file. Add ActiveProblems snapshot test to ClinicalInfo.test.tsx. Update tests.
